+++ src/main/java/com/purej/vminspect/html/SystemMainView.java
@@ -125,7 +125,7 @@ public final class SystemMainView extends AbstractHtmlView {
   }
 
   private static String toShowLinkWithHiddenDiv(String divId, String txt) {
-    return showHideLink(divId, "Show") + "<br/>" + "<div id='" + divId + "' style='display: none;'>" + htmlEncode(txt) + "</div>";
+    return showHideLink(divId, "Show") + "<br/>" + "<div id='" + divId + "' class='hidden'>" + htmlEncode(txt) + "</div>";
   }
 
   private static String toMemoryBar(MemoryData memory) {